refactor: move hex_to_root out of steps.rs into mod.rs
hex_to_root is a generic hex->Root primitive, not a step/YAML type -- and the
`pub(crate) use steps::hex_to_root` re-export was the tell that the module
already treated it as module-level, not a steps concern. Define it in mod.rs
directly and drop the re-export indirection. beacon_header_matches stays in
steps.rs (it reads a HeaderCheck, a steps type).
