+++ src/test_utils/README.md
@@ -20,9 +20,9 @@ the light client against the official vectors with no network or beacon node.
 |------|----------------|
 | `loader.rs` | `SpecTestLoader` — the entry point; reads fixture files and returns typed objects |
 | `raw_ssz.rs` | `Raw*` SSZ structs + the `raw_*_to_pub` raw→production converters |
-| `steps.rs` | YAML fixture types (`meta.yaml` / `steps.yaml`) + `hex_to_root` |
+| `steps.rs` | YAML fixture types (`meta.yaml` / `steps.yaml`) + `beacon_header_matches` |
 | `fork.rs` | `MinimalPresetFork` — the minimal-preset fork tag and its `ChainSpec` |
-| `mod.rs` | module wiring / re-exports |
+| `mod.rs` | module wiring / re-exports + `hex_to_root` |
 
 ## Fixtures
